backend/requirements.txt
django==4.2.20
djangorestframework==3.16.0
django-cors-headers==4.7.0
mypy==1.15.0
django-stubs==5.1.3
djangorestframework-stubs==3.15.3 